\name{LAMGrid-class}
\Rdversion{1.1}
\docType{class}
\alias{LAMGrid-class}

\title{Class \code{"LAMGrid"}}
\description{
A class defining a linear grid for LAM
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("LAMGrid", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{nlong}:}{Object of class \code{"integer"} ~~ }
    \item{\code{center}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{lat}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{long}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{Grid}"}, directly.
}
\section{Methods}{
No methods defined with class "LAMGrid" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("LAMGrid")
}
\keyword{classes}
