\name{mffield-package}
\alias{mffield-package}
\alias{mffield}
\docType{package}
\title{
\packageTitle{mffield}
}
\description{
\packageDescription{mffield}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mffield}
\packageIndices{mffield}
Functions and classes for NWP fields defined on grids, levels and for forecast, observed
 or climatological times. Provide basic functionality like interpolation and vertical
 cross-section. Support reduced Gaussian grid and Lambert LAM grid.
}
\author{
\packageAuthor{mffield}

Maintainer: \packageMaintainer{mffield}
}
\references{
No references
}
\keyword{
package
NWP
}
\seealso{
Package mfnode for arsing and viewing information from NODE file(s)
\code{\link[<pkg>:<pkg>-package]{mfnode}}
}
\examples{
No examples
}
